Addressed SAIL-35 Remove the data from Sailthru when it's been deleted in Magento
- Refactoring of send requests with product data to Sailthru
