+++ Plugin/ProductIntercept.php
@@ -2,29 +2,50 @@
 
 namespace Sailthru\MageSail\Plugin;
 
-use Magento\Catalog\Model\Product;
-use Sailthru\MageSail\Helper\Product as ProductHelper;
+use Sailthru\MageSail\Helper\ClientManager;
+use Sailthru\MageSail\Helper\Product\PostContent as PostContentHelper;
+use Magento\Store\Model\StoreManagerInterface;
 
 class ProductIntercept
 {
     /**
-     * @var ProductHelper
+     * @var ClientManager
      */
-    protected $productHelper;
+    protected $clientManager;
 
     /**
-     * @param ProductHelper $productHelper
+     * @var PostContentHelper
+     */
+    protected $postContentHelper;
+
+    /**
+     * @var StoreManagerInterface
+     */
+    protected $storeManager;
+
+    /**
+     * @param ClientManager $clientManager
+     * @param PostContentHelper $postContentHelper
+     * @param StoreManagerInterface $storeManager
      */
     public function __construct(
-        ProductHelper $productHelper
+        ClientManager $clientManager,
+        PostContentHelper $postContentHelper,
+        StoreManagerInterface $storeManager
     ) {
-        $this->productHelper = $productHelper;
+        $this->clientManager = $clientManager;
+        $this->postContentHelper = $postContentHelper;
+        $this->storeManager = $storeManager;
     }
 
     public function afterAfterSave(Product $subject, Product $result)
     {
         try {
-            $this->productHelper->postContent($result);
+            if ($this->storeManager->getStore($result->getStoreId())->getCode() == 'admin') {
+                $this->postContentHelper->sendMultipleRequests($result);
+            } else {
+                $this->postContentHelper->sendRequest($result);
+            }
         } catch (\Throwable $t) {
             $client = $this->clientManager->getClient($result->getStoreId());
             $client->logger(__('Error of send product data to Sailthru'));
